WIP: noop(refactor) fully moves video logic to interface
